Move twig extensions to dedicated folder
